Adapt XML parser to new data format
The records now encode the availability as an IntArray of size 12 to
quickly check for availability for a given month as an index. Also, new
data is recorded, like shadow size for fish and location.

In addition, the availability info has been removed from the list view,
as it occupies too much space. This info will be available in the future
in a calendar view, a detailed view of the collectible, or by using
filters.