First implementation of filters
Allow a recordset to be filtered into a subset of it by setting a custom
condition on records.

Add a floating action button on the bottom right of the screen for
filter functionality. When clicked, a submenu of action buttons will
open. For now there are two options: filter records by availability on
the current month (grabbed from the device), or clear applied filters,
but in the future it can hold more options like filter by
name/month/etc.