Allow sorting of records
Allow sorting of records by name or price, in ascending or descending
order. The sorting option has been added to the floating action button
submenu, and when clicked will open a new dialog to select the desired
sorting configuration.

For now, sorting, as well as filters, are not persistent between
recordsets, meaning that changing to a different recordset (e.g. from
fish to bugs) will clear all filters and sorting options.