Make filters/sorting cumulative
Filters and sorting are now applied to the currently shown recordset.
This means that if we sort by price, and then filter by month, we will
get the filtered results already sorted. If we then filter by name, we
will filter on the already filtered results instead of the entire
recordset.

To clear all applied filters and sortings, for now one can click on the
'Clear all filters' button