Optimize asset loading and properly update layout on game change
With this commit, the reloading of game assets is optimized so that
whenever the dataset changes, only the Adapter's internal dataset is
changed without having to recreate and reinstanciate it.

This commit also fixes #13 which wouldn't reload the layout UI when
changing the game from the burger menu, this is fixed simply
implementing `onTabReselected` of `tabLayout`.