Correctly hide submenu FABs when not active
This fixes the bug where taps on where a submenu button would be are not
registered, even if the submenu is not open. Even though the menu
buttons were invisible and disabled, they weren't really gone from the
layout, and thus were blocking the taps.

With this change, the buttons are correctly hidden, fixing this bug.

Fixes #21