Filter and sorting overhaul
Rework how sorting and filters are interacted with. Instead of using a
submenu within the FAB in the main activity, now filters and sortings
are in a drawer that pops up from the right when the user clicks on the
filter button on the toolbar menu.

Filters have a checkbox view, and will apply as soon as you
check/uncheck them. They are stackable, so you can filter by more than
one criterion at a time.

Sortings are shown on a list, with the currently applied sorting having
an arrow image on the left, pointing upward if the sort is ascending, or
downward if it is descending. Clicking on the active sorting will switch
between ascending/descending order, and clicking on a different sorting
will change the sorting to sort by that new item, in the same order as
the one that was selected before. To avoid confusion, only one sorting
is allowed at a time. Sorting will also apply as soon as some sort item
is clicked.

On another note, filters and sorting have been made persistent between
tab and game changes.

On the technical side, how filters/sorting work has also been
refactored, moving the functionality out of CollectionActivity (view),
and creating a controller to handle each of them, which communicates
with the model (Recordset).

As a result of this commit, the FAB has been rendered useless for now,
even though it will maintain some functionality like hiding/reappearing
when the user scrolls down/up respectively, but clicking it will do
nothing. It has been left there regardless, as in a future PR search
functionality will be added to it.