Add search feature
Now it is possible to filter records searching by name. The search
functionality can be accessed through a new icon on the toolbar, which
opens an edit text in which the user can type. The contents entered will
be used to filter records by name, ignoring case, and the recordset will
update as the user types.

Like other filters, the search query is persistent between tabs/games.

This commit also gets rid of the FAB completely, as it has no use for
now.