Use fragments to easily change views
This commit moves existing views (CollectionActivity, AboutActivity)
into fragments in order to keep the change between views consistent
(burger menu across most views) and smooth.

This allows us the possibility to add new features / views easily while
keeping the burger menu visible across all of them.

Some changes had to be made to FilterManager and SortManager and the way
they work in order to acommodate for the fact that the views are
fractured now.