Prepare for internationalization
To prepare for translating the app into different languages, critter
names have been extracted into string resources from the assets XMLs,
which now have the name of the string resource in the <name> tags
instead of the actual name string.

The same has been done for location strings, but not for time strings,
as they do not require a translation and they will be encoded
differently in the future to allow filtering by time.