Update resources with proper size images
Previous images were limited to 64x64 and were being upscaled on bigger
screens, leading to blurry images.

With this commit, the version of the image at full size is used and then
downscaled to a maximum of 80dp which should be plenty for both small
and big screens without any blurriness.