Add fossils and basic european l18n
With this commit, fossils are added for both ACNH and ACNL, this entails
multiple changes such as sub-menus for every game, critters and fossils
for now.

This commit also introduces translates for, at the very least, the name
of critters / fossils in german, spanish, french, italian and dutch,
more languages are planned in the future.